requirements.txt
altgraph==0.16.1
cycler==0.10.0
Django==2.2.6
django-redis==4.10.0
django-tinymce==3.0.2
future==0.18.0
kiwisolver==1.2.0
matplotlib==3.0.3
mysqlclient==1.4.4
numpy==1.18.4
pefile==2019.4.18
Pillow==6.2.0
PyInstaller==3.5
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
pytz==2019.2
pywin32-ctypes==0.2.0
redis==3.3.8
six==1.14.0
sqlparse==0.3.0
